requirements.txt
python-dotenv
openai
google-cloud-speech
pyaudio
google-generativeai
python-dotenv
firebase-admin
